Multi-Profiles: Update cached user name
Initially ProfileInfoCache didn't store the sync user name. Later when we added support for this field we only updated it if the sync user name value changed.

This meant that for existing multi-profiles users we didn't cache their sync user name.

To fix this I'm adding code to make sure that the sync user name is up to date when we create a new ProfileSyncService.

BUG=99531
TEST=


Review URL: http://codereview.chromium.org/8498026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@109119 0039d316-1c4b-4281-b951-d872f2087c98