UserPolicyCache only becomes ready after policy has been fetched.
This behavior is only forced on enterprise devices, and for logged in users whose username matches the enrollment domain.

This requires waiting for a couple of things:
- oauth token fetching
- dm server register token fetching
- user policy fetching

The cache resumes initialization too if any of these steps fails. Fetching of OAuth tokens usually happens after Profile creation, but in this case it will be done earlier so that the policy fetching can proceed before the Profile is fully created.

This fixes the races between policy and sync, startup pages, etc.

BUG=chromium-os:17698
TEST=Logins of enterprise users on enrolled ChromeOS devices waits for a policy fetch before proceeding.

Review URL: http://codereview.chromium.org/8499021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@110108 0039d316-1c4b-4281-b951-d872f2087c98