Open NaCl IRT file only once at startup
This changes the NaCl startup dance so that the IRT file is opened
just once at browser startup.  That file descriptor is kept around and
passed repeatedly to each NaCl process launched.  This ensures that
when autoupdate replaces the file on disk with a new version, we
continue to use the original file that corresponds to the old browser
version that's still running.

We also eliminate the cases for not having an IRT file, which is now a
hard error (i.e. prevents NaCl launches).  It's been a hard
requirement for NaCl that the IRT be available since Chromium 14.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=1772
TEST= hand-tested in Chromium build on Linux, Mac, and Windows

R=brettw@chromium.org

Review URL: http://codereview.chromium.org/8397001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@110136 0039d316-1c4b-4281-b951-d872f2087c98