Cloud print connector policy.
This implements the policy inside the browser process, shutting down the cloud
print connector if the policy is set to disabled. This isn't a complete
solution, as the browser needs to be running or be launched for the policy to
take effect. (It will take a lot more refactoring for the service process to
use the policy code). The hole without the refactoring is that if you enable
the connector, quit Chromium, and set the policy, the connector will stay alive
until the next launch of Chromium.

The browser process checks the policy on startup, and listens for it changing
as long as it is up. You can sit on the Under the Hood page and watch the
button change state on Windows as you fiddle with the policy.

BUG=59769
TEST=Set Cloud Print Proxy policy to disabled, bring up browser, Options/Under
the Hood - Sign in to Google Cloud print will be disabled.  Unset policy, wait,
button becomes active.  Log in to cloud print.  Quit Chromium, note service
process hangs around for more than a minute.  Set policy, launch and quit
Chromium. Note that the service process quits within a minute.


Review URL: http://codereview.chromium.org/8438020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@110332 0039d316-1c4b-4281-b951-d872f2087c98