Replace FileManagerUtil class with namespace file_manager_util.
file_manager_util namespace is consistent with other utils in the codebase (platform_util, mime_util, file_util), and is the preferred way in the google style guide to do this kind of thing. The FileManagerUtil class is never initialized and has no state.

BUG=None, cleanup
TEST=Compiles, tests pass.
Review URL: http://codereview.chromium.org/8574054

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@110600 0039d316-1c4b-4281-b951-d872f2087c98