Add gamepad hardware data fetcher, and higher level thread container
DataFetcher is the hardware and platform-specific data getter that talks to OS
level interfaces for each platform and fill in a structure provided to it.
Currently only Windows is here.

Provider is a manager for the background thread that the DataFetcher runs on,
and interacts with the rest of browser.

This is part of a larger patch, the remainder of which is at
http://codereview.chromium.org/8345027/

BUG=79050


Review URL: http://codereview.chromium.org/8568029

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@111189 0039d316-1c4b-4281-b951-d872f2087c98