Work around a bug in gyp triggered by webkit_appcache.gypi.
The bug seems to be showing up on the chromium build.webkit.org bots
where we're generating two different versions of webkit_appcache.target.mk
and the second one is overwriting the first, but is being considered
to not be "part of all". As a result, DumpRenderTree is no longer being
built by "make all".

TBR=tony@chromium.org

BUG=
TEST=

Review URL: http://codereview.chromium.org/8667009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@111325 0039d316-1c4b-4281-b951-d872f2087c98