Add a mechanism to launch the utility process from the zygote on Linux.
Allow the users of the utility process to specify whether the zygote
should be used on Linux. This will allow sandboxing the uses of the
utility process that don't do FS access.

This is the first step to get the utility process sandboxed
on Linux. Since most of the uses of the utility process
don't do file access, launching all of those from the zygote
will simplify sandboxing the one that does: extension
unpacking.

Subsequent CLs will port the uses of the utility process
that don't do file access to launch from the zygote.

BUG=93109
TEST=No functional changes yet, but test that nothing broke by
running browser_tests::ExtensionWebstorePrivateApiTest.* and also
logging into Chromium and installing an extension from the Web Store.


Review URL: http://codereview.chromium.org/8849007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@113914 0039d316-1c4b-4281-b951-d872f2087c98