Make ui_tests fail faster if all tests fail.
UITest::CreateProxyLauncher() creates a proxy launcher that severes
the IPC connection to the browser process on AutomationProxy::Send().
If a bug makes the browser misbehave on startup, the IPC connection
is dropped immediately, which means that
ProxyLauncher::WaitForBrowserProcessToQuit() will always timeout.
Fix this by assuming that waiting for shutdown will always fail if
the IPC connection was prematurely dropped.

BUG=106517
TEST=Change a test to timeout. The runner now takes 2*action_timeout (90s) instead 3*action_timeout (135s) to realize that.


Review URL: http://codereview.chromium.org/8907004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@115017 0039d316-1c4b-4281-b951-d872f2087c98