Switch to PostTaskAndReply for AccessTokenStore.
This removes the only usage of the CancelableRequest class in
content/, and is in preparation for moving the class back to
chrome/browser.

The ability to cancel requests to the AccessTokenStore was never used in production, so the interface now changes to a simpler non-cancelable one, and the browsertest GeolocationAccessTokenStoreTest.CancelRequest which tested the functionality is removed.

BUG=98716

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=114992
Reverted (ThreadChecker DCHECK on debug bots): r114997

Review URL: http://codereview.chromium.org/8996006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@115325 0039d316-1c4b-4281-b951-d872f2087c98