On Windows, D3D devices used to present the compositor's output are recycled.
This is because some drivers seem to crash when a D3D query object is released. To avoid crashing, the D3D device and associated query are stored in a cache to be used by future tabs. The device's swap chain is resized to 1 x 1 during this time so that memory is not wasted.

In order for the D3D objects to outlive the AcceleratedSurface object, I broke them out into another class.

I considered cleaning up the D3D objects at the time where the browser process terminates. That would result in a crash on browser termination, which should not impact the user. However, I went with the former approach because it is simpler. The D3D device will likely crash if it is destroyed on the main thread, the thread where the LazyInstances are cleaned up. By the time the LazyInstances are cleaned up, the only thread on which the D3D device can correctly be destroyed can likely no longer do so.

BUG=108317

Review URL: http://codereview.chromium.org/9087022

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@116520 0039d316-1c4b-4281-b951-d872f2087c98