Remove a few trivial tab_contents.h includes in chrome that I missed earlier.
I've switched RenderViewHostTestHarness::controller() to return the interface instead of the concrete class. Only two tests needed NavigationControllerImpl, so it's much easier to switch those to cast, and to leave the interface object as the return value as many tests in chrome use it and RVHTH will have to be exposed to chrome.

BUG=98716
Review URL: http://codereview.chromium.org/9123002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@116681 0039d316-1c4b-4281-b951-d872f2087c98