Fix NTP Crash when Moving CWS Icon.
Fixing the code to call the code to initialize the page ordinal map, even if
there are no extensions. This ensures that the chrome web store is properly
loaded into the map.

BUG=110269
TEST=When the only icon on the NTP is the chrome web store, the user should be
able to move it to a new page and not have chrome crash.


Review URL: http://codereview.chromium.org/9231012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@117898 0039d316-1c4b-4281-b951-d872f2087c98