Added explanatory dialog for enterprise users during OOBE.
This is only shown when auto-enrollment is performed and doesn't succeed
automatically at first. It explains the user what he has to do, and gives a
(buried) option to opt out of enterprise enrollment.

This feature is behind a flag, and is still disabled by default.

BUG=chromium-os:23063
TEST=Login with an invalid (e.g. @gmail.com) account at the first sign-in screen during OOBE. Auto-enrollment should fail and show an explanatory dialog.


Review URL: http://codereview.chromium.org/9022049

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@118087 0039d316-1c4b-4281-b951-d872f2087c98