sync: Make ProfileSyncService a ProfileKeyedService.
Create ProfileSyncService using a factory, and transfer ownership of the components factory
to ProfileSyncService. I deliberately left GetProfileSyncService to
keep this patch focused on the actual change and more manageable in size. It now just wraps
calls to the factory, and I'll change the rest of the callers in a follow up (there are a lot).

BUG=93922
TEST=Sync still works / can be set up / starts properly after browser restart / shuts down.

Review URL: https://chromiumcodereview.appspot.com/9232011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@118474 0039d316-1c4b-4281-b951-d872f2087c98