Fix flashing window when deleting profile
Deleting a profile with the sync promo open would cause the delete profile's browser window to temporarily re-open.

The problem was that the sync promo was attempting to navigate to a page when the promo was closed. Fix was to disable the navigation if the browser window was being closed.

BUG=108801
TEST=


Review URL: http://codereview.chromium.org/9193004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@118710 0039d316-1c4b-4281-b951-d872f2087c98