Separate the generated date from xkeyboard.cc for ease of udpate.
Currently we need to be careful not to delete the start and the end marker lines so that gen_keyboard_overlay_data.py work properly. This change move the generated part of xkeyboard.cc to a different file in order to make it easy to update the data and xkeyboard.cc.

BUG=None
TEST=unit_tests --gtest_filter="XKeyboardTest.*" passes

Review URL: http://codereview.chromium.org/9289044

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@119293 0039d316-1c4b-4281-b951-d872f2087c98