Avoid trying to query ScreenInfo without a valid host_window_
It's possible for RenderWidget::screenInfo() to be called before the host_window_ id is set,
either because we haven't received the IPC yet with the id or because the native widget hasn't
been created yet. The implementations of this all fall over in different platform-specific ways
without the parameter, and since this is a synchronous IPC it seems better to avoid the round-trip
if we know it won't be successful.

This returns a WebScreenInfo in the default state (zero'd out). It'd be better to have a way to signal
an error but that would require a WebKit API change.

TEST=


Review URL: http://codereview.chromium.org/9169103

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@119520 0039d316-1c4b-4281-b951-d872f2087c98