uber extensions
change trash to be a button so that clicking it doesn't focus it (as I recently discovered).

the problem described in the bug stems from the fact that clicking the div focuses it, and a focused trash icon is visible and open (to support keyboard usage).

BUG=112496
TEST=see bug

Review URL: https://chromiumcodereview.appspot.com/9323052

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@120674 0039d316-1c4b-4281-b951-d872f2087c98