Framing for a DomStorage backend that does not depend on in-process-webkit. Only partial unittests coverage in this CL, more to come.
These classes aren't used yet.

Classes intended to be used by an embedder are DomStorageContext,DomStorageHost,
and DomStorageSession. The other classes are for internal consumption.

In general terms, the DomStorage object relationships are...
* Contexts (per-profile) own Namespaces which own Areas which share Maps.
* Hosts(per-renderer) refer to Namespaces and Areas open in it's renderer.
* Sessions (per-tab) cause the creation and deletion of session Namespaces.

Session Namespaces are cloned by initially making a shallow copy of
all contained Areas, the shallow copies refer to the same refcounted Map,
and does a deep copy-on-write if needed.

BUG=106763
Review URL: https://chromiumcodereview.appspot.com/9146025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@120687 0039d316-1c4b-4281-b951-d872f2087c98