chromeos: bluetooth: tie Proxy lifetime to object, not observer
Change the lifetime of Object Proxies to the remote object, not any
Observer; one of the problems with the previous method was adding
a second observer would treat every new attempt to add an observer
as a new proxy so reconnect signals, etc.

The other main problem is that you couldn't call methods unless there
was an observer that had created the object proxy.

Originally landed in I9317012, reverted in I9316113; unfortunately CQ
was letting it through without the fixes to the Mock code.

BUG=chromium-os:22086
TEST=verified methods still work, built tests with USE=aura

Change-Id: Icc86c03d00dabc27f2b31f9549da3a163f4b18e8


Review URL: http://codereview.chromium.org/9309122

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@120853 0039d316-1c4b-4281-b951-d872f2087c98