Avoid overread in RenderWidgetHostViewWin::OnPaint
GetRegionData may return 0 to indicate error. This call is used both to
calculate the required buffer size and to fill the buffer with region
data. In the first case, if not checked, a zero-length buffer may be
allocated. In the second case, the buffer content is undefined. In either
case, we should not depend on the buffer content.

BUG=110176
TEST=Long tail crasher in gfx::Rect::Rect, called from
RenderWidgetHostViewWin::OnPaint.


Review URL: http://codereview.chromium.org/9316056

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@120916 0039d316-1c4b-4281-b951-d872f2087c98