Restores Fullscreen mode during session restore on Mac.  Serializes and deserializes fullscreen state during session restore.  Also, fixes issue where entering fullscreen mode is interupted it was leaving the window tab strip layout in a weird state on Lion.
There are two issues I'll address in a follow on patch.  (1) On
Lion "real" session restore suppresses the animation of entering
fullscreen.  (2) On Lion, if multiple windows are in fullscreen
mode, they are assigned to different spaces.  Currently we restore
only on window to fullscreen on Lion due to a system issue that
fails to sequence the subsequent fullscreen animations.

BUG=94611
TEST=BrowserWindowCocoaTest.TestFullscreen
R=rohit@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9323070

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@121146 0039d316-1c4b-4281-b951-d872f2087c98