Refactoring TestMediaStreamClient
As part of a major effor to add LayoutTest capabilities to the WebRTC feature, I need to move WebUserMediaClientMock from WebKit/Tools/DumpRenderTree/chromium/ to WebKit/Source/WebKit/chromium/. This makes the class unable to inherit from MediaStreamUtil, and fixing that is what this patch is about. I am moving functionality from WebUserMediaClientMock to TestMediaStreamClient, which is a better place for it.

This is patch #1 of 3 (chromium - webkit - chromium)

BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/9380009
Patch from Tommy Widenflycht <tommyw@google.com>.

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@122136 0039d316-1c4b-4281-b951-d872f2087c98